requirements.txt
addict==2.4.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.2
fonttools==4.48.1
GPUtil==1.4.0
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipython==8.12.3
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lightning-utilities==0.10.1
MarkupSafe==2.1.5
matplotlib==3.7.4
matplotlib-inline==0.1.6
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
open3d==0.18.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.18.0
plyfile==1.0.3
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
pyquaternion==0.9.9
python-dateutil==2.8.2
pytorch-msssim==1.0.0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
retrying==1.3.4
rpds-py==0.18.0
scikit-image==0.19.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
threadpoolctl==3.3.0
tifffile==2023.7.10
torch==1.12.1+cu113
torchmetrics==1.3.1
torchvision==0.13.1+cu113
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.10
zipp==3.17.0
